polys modularised, still not better
